solve circular dep
