+++ hostxdc/integration_domain_failover_test.go
@@ -44,6 +44,7 @@ import (
 	workflow "github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/cache"
+	"github.com/uber/cadence/environment"
 	"github.com/uber/cadence/host"
 	"gopkg.in/yaml.v2"
 )
@@ -103,7 +104,7 @@ func (s *integrationClustersTestSuite) SetupSuite() {
 	if *host.TestClusterConfigFile != "" {
 		fileName = *host.TestClusterConfigFile
 	}
-	host.SetupEnv()
+	environment.SetupEnv()
 
 	confContent, err := ioutil.ReadFile(fileName)
 	s.Require().NoError(err)
