+++ tools/cassandra/setupTask_test.go
@@ -26,11 +26,12 @@ import (
 	"testing"
 	"time"
 
+	"github.com/uber/cadence/environment"
+
 	"github.com/gocql/gocql"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"github.com/uber-common/bark"
-	"github.com/uber/cadence/host"
 
 	"io/ioutil"
 	"os"
@@ -64,7 +65,7 @@ func (s *SetupSchemaTestSuite) SetupSuite() {
 	s.rand = rand.New(rand.NewSource(time.Now().UnixNano()))
 	s.keyspace = fmt.Sprintf("setup_schema_test_%v", s.rand.Int63())
 
-	client, err := newCQLClient(host.GetCassandraAddress(), defaultCassandraPort, "", "", "system", defaultTimeout)
+	client, err := newCQLClient(environment.GetCassandraAddress(), defaultCassandraPort, "", "", "system", defaultTimeout)
 	if err != nil {
 		s.log.Fatal("Error creating CQLClient")
 	}
@@ -90,7 +91,7 @@ func (s *SetupSchemaTestSuite) TestCreateKeyspace() {
 
 func (s *SetupSchemaTestSuite) TestSetupSchema() {
 
-	client, err := newCQLClient(host.GetCassandraAddress(), defaultCassandraPort, "", "", s.keyspace, defaultTimeout)
+	client, err := newCQLClient(environment.GetCassandraAddress(), defaultCassandraPort, "", "", s.keyspace, defaultTimeout)
 	s.Nil(err)
 
 	// test command fails without required arguments
