+++ tools/cassandra/updateTask_test.go
@@ -33,7 +33,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"github.com/uber-common/bark"
-	"github.com/uber/cadence/host"
+	"github.com/uber/cadence/environment"
 )
 
 type (
@@ -62,7 +62,7 @@ func (s *UpdateSchemaTestSuite) SetupSuite() {
 	s.rand = rand.New(rand.NewSource(time.Now().UnixNano()))
 	s.keyspace = fmt.Sprintf("update_schema_test_%v", s.rand.Int63())
 
-	client, err := newCQLClient(host.GetCassandraAddress(), defaultCassandraPort, "", "", "system", defaultTimeout)
+	client, err := newCQLClient(environment.GetCassandraAddress(), defaultCassandraPort, "", "", "system", defaultTimeout)
 	if err != nil {
 		s.log.Fatal("Error creating CQLClient")
 	}
@@ -82,7 +82,7 @@ func (s *UpdateSchemaTestSuite) TearDownSuite() {
 
 func (s *UpdateSchemaTestSuite) TestUpdateSchema() {
 
-	client, err := newCQLClient(host.GetCassandraAddress(), defaultCassandraPort, "", "", s.keyspace, defaultTimeout)
+	client, err := newCQLClient(environment.GetCassandraAddress(), defaultCassandraPort, "", "", s.keyspace, defaultTimeout)
 	s.Nil(err)
 	defer client.Close()
 
@@ -119,7 +119,7 @@ func (s *UpdateSchemaTestSuite) TestUpdateSchema() {
 
 func (s *UpdateSchemaTestSuite) TestDryrun() {
 
-	client, err := newCQLClient(host.GetCassandraAddress(), defaultCassandraPort, "", "", s.keyspace, defaultTimeout)
+	client, err := newCQLClient(environment.GetCassandraAddress(), defaultCassandraPort, "", "", s.keyspace, defaultTimeout)
 	s.Nil(err)
 	defer client.Close()
